+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,5 +1,7 @@
 package roomescape.domain;
 
+import roomescape.exception.BlankReservationException;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 
@@ -37,19 +39,19 @@ public class Reservation {
 
     private static void validateName(String name) {
         if (name == null || name.isBlank()) {
-            throw new IllegalArgumentException("이름을 입력해주세요");
+            throw new BlankReservationException("이름을 입력해주세요");
         }
     }
 
     private static void validateDate(LocalDate date) {
         if (date == null) {
-            throw new IllegalArgumentException("날짜를 선택해주세요");
+            throw new BlankReservationException("날짜를 선택해주세요");
         }
     }
 
     private static void validateTime(LocalTime time) {
         if (time == null) {
-            throw new IllegalArgumentException("시간을 선택해주세요");
+            throw new BlankReservationException("시간을 선택해주세요");
         }
     }
 }
