+++ src/main/java/roomescape/exception/BlankReservationException.java
@@ -1,0 +1,7 @@
+package roomescape.exception;
+
+public class BlankReservationException extends RuntimeException{
+    public BlankReservationException(String message) {
+        super(message);
+    }
+}
