+++ src/main/java/roomescape/exception/GlobalExceptionHandler.java
@@ -1,0 +1,18 @@
+package roomescape.exception;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class GlobalExceptionHandler {
+    @ExceptionHandler(NotFoundReservationException.class)
+    public ResponseEntity<Void> handleNotFoundReservation(NotFoundReservationException e) {
+        return ResponseEntity.notFound().build();
+    }
+
+    @ExceptionHandler(BlankReservationException.class)
+    public ResponseEntity<Void> handleIllegalArgument(IllegalArgumentException e) {
+        return ResponseEntity.badRequest().build();
+    }
+}
